Moved group and personal card delete and update resources to ProfileResource
